myshop/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cssselect2==0.7.0
Deprecated==1.2.18
distro==1.9.0
Django==5.1.5
django-celery-beat==2.7.0
django-csp==4.0
django-debug-toolbar==5.1.0
django-filter==25.1
django-localflavor==4.0
django-parler==2.3
django-redis==5.4.0
django-rosetta==0.10.1
django-timezone-field==7.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
flake8==7.1.2
flower==2.0.1
fonttools==4.56.0
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jst-parler==1.0.0
jst-parler-rest==1.0.1
kombu==5.5.2
mccabe==0.7.0
netaddr==1.3.0
ngrok==1.4.0
packaging==24.2
parler==1.0.1
phonenumbers==8.13.53
pillow==11.1.0
polib==1.2.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
pyflakes==3.2.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
stripe==9.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
wrapt==1.17.2
yookassa==3.5.0
zopfli==0.2.3.post1
